requirements.txt
pygame-ce
fastapi
uvicorn[standard]
pydantic
torch
torchvision
numpy
matplotlib
tensorboard